Initial rebase

+++ code/datums/diseases/advance/symptoms/weight.dm
@@ -19,21 +19,33 @@ Bonus
 /datum/symptom/weight_loss
 
 	name = "Weight Loss"
-	stealth = -3
-	resistance = -2
+	desc = "The virus mutates the host's metabolism, making it almost unable to gain nutrition from food."
+	stealth = -2
+	resistance = 2
 	stage_speed = -2
 	transmittable = -2
 	level = 3
-	severity = 1
+	severity = 3
+	base_message_chance = 100
+	symptom_delay_min = 15
+	symptom_delay_max = 45
+	threshold_desc = "<b>Stealth 4:</b> The symptom is less noticeable."
+
+/datum/symptom/weight_loss/Start(datum/disease/advance/A)
+	if(!..())
+		return
+	if(A.properties["stealth"] >= 4) //warn less often
+		base_message_chance = 25
 
 /datum/symptom/weight_loss/Activate(datum/disease/advance/A)
-	..()
-	if(prob(SYMPTOM_ACTIVATION_PROB))
-		var/mob/living/M = A.affected_mob
-		switch(A.stage)
-			if(1, 2, 3, 4)
+	if(!..())
+		return
+	var/mob/living/M = A.affected_mob
+	switch(A.stage)
+		if(1, 2, 3, 4)
+			if(prob(base_message_chance))
 				to_chat(M, "<span class='warning'>[pick("You feel hungry.", "You crave for food.")]</span>")
-			else
-				to_chat(M, "<span class='warning'><i>[pick("So hungry...", "You'd kill someone for a bite of food...", "Hunger cramps seize you...")]</i></span>")
-				M.overeatduration = max(M.overeatduration - 100, 0)
-				M.nutrition = max(M.nutrition - 100, 0)
+		else
+			to_chat(M, "<span class='warning'><i>[pick("So hungry...", "You'd kill someone for a bite of food...", "Hunger cramps seize you...")]</i></span>")
+			M.overeatduration = max(M.overeatduration - 100, 0)
+			M.adjust_nutrition(-100)
